Updated sc.py with some minor bugs
Reason for modifying it:
When I ran the program and called cancel_shutdown function only it worked well and when I ran this program second time and called the same function second time then I got an output that "Unable to abort the system shutdown because no shutdown was in progress.(1116)" after that I got another output from the program "Shutdown has been cancelled!". So the output with numeric code was not from the program but from the windows command line. Now what I actually did is that whenever cancel_shutdown function will be called under first condition it will check with the error code that whether a system shutdown has already scheduled or not and the program under that condition will behave accordingly, the same has been implemented under the shutdown function's first condition[Actually I'm a windows user, so I amended the changes under the "windows" which is set as a condition].